added pattern test cases
