+++ pkg/detectors/anypoint/anypoint.go
@@ -3,9 +3,10 @@ package anypoint
 import (
 	"context"
 	"fmt"
-	regexp "github.com/wasilibs/go-re2"
 	"net/http"
 	"strings"
+
+	regexp "github.com/wasilibs/go-re2"
 
 	"github.com/trufflesecurity/trufflehog/v3/pkg/common"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors"
@@ -50,6 +51,11 @@ func (s Scanner) FromData(ctx context.Context, verify bool, data []byte) (result
 				continue
 			}
 			orgRes := strings.TrimSpace(orgMatch[1])
+
+			// regex for both key and org are same, so to avoid same string processing
+			if resMatch == orgRes {
+				continue
+			}
 
 			s1 := detectors.Result{
 				DetectorType: detectorspb.DetectorType_Anypoint,
